+++ src/theme.js
@@ -1,189 +1,220 @@
 const chroma = require("chroma-js");
 const { getColors } = require("./colors");
 
-// Choosing colors
-// There are two ways to define what color is used:
+// Choosing colors from primer/primitives
+// There are multiple ways to define what color is used:
 
-// 1. Color scale from primer/primitives.
-//    e.g. "textLink.foreground": themes({ light: scale.blue[5], dark: scale.blue[2], dimmed: scale.blue[3] }),
-// 2. Hex value: All themes will use this scale. Only use for exceptions
-//    e.g. "textLink.foreground": "#fff",
+// 1. Global variable
+//    e.g. "textLink.foreground": color.fg.default,
+// 2. Color scale
+//    e.g. "textLink.foreground": scale.blue[5],
+// 3. Per theme. Useful when a certain theme needs an exception
+//    e.g. "textLink.foreground": themes({ light: scale.blue[5], light_high_contrast: scale.blue[5], light_colorblind: scale.blue[5], dark: scale.blue[2], dark_high_contrast: scale.blue[3], dark_colorblind: scale.blue[2], dark_dimmed: scale.blue[3] }),
 
 function getTheme({ theme, name }) {
-  // Usage: themes({ light: "lightblue", dark: "darkblue", dimmed: "royalblue" })
-  const themes = (options) => options[theme];
-  // Usage: color.text.primary
-  const color = getColors(theme);
-  // Usage: scale.blue[6]
-  const scale = color.scale;
+
+  const themes = (options) => options[theme]; // Usage: themes({ light: "lightblue", light_high_contrast: "lightblue", light_colorblind: "lightblue", dark: "darkblue", dark_high_contrast: "darkblue", dark_colorblind: "darkblue", dark_dimmed: "royalblue" })
+  const rawColors = getColors(theme)
+  const color = changeColorToHexAlphas(rawColors)
+  const scale = color.scale; // Usage: scale.blue[6]
+
+  const onlyDark = (color) => {
+    return themes({ dark: color, dark_high_contrast: color, dark_colorblind: color, dark_dimmed: color })
+  }
+
+  const onlyDarkHighContrast = (color) => {
+    return themes({ dark_high_contrast: color })
+  }
+
+  const lightDark = (light, dark) => {
+    return themes({ light: light, light_high_contrast: light, light_colorblind: light, dark: dark, dark_high_contrast: dark, dark_colorblind: dark, dark_dimmed: dark })
+  }
+
+  const alpha = (color, alpha) => {
+    return chroma(color).alpha(alpha).hex()
+  }
 
   return {
     name: name,
     colors: {
-      focusBorder          : color.state.focus.border,
-      foreground           : color.text.primary,
-      descriptionForeground: color.text.tertiary,
-      errorForeground      : color.text.danger,
+      focusBorder          : color.accent.emphasis,
+      foreground           : color.fg.default,
+      descriptionForeground: color.fg.muted,
+      errorForeground      : color.danger.fg,
 
-      "textLink.foreground"      : color.text.link,
-      "textLink.activeForeground": color.text.link,
-      "textBlockQuote.background": color.bg.canvasInset,
-      "textBlockQuote.border"    : color.markdown.blockquoteBorder,
-      "textCodeBlock.background" : chroma(color.markdown.codeBg).hex(),
-      "textPreformat.foreground" : color.text.secondary,
-      "textSeparator.foreground" : color.border.secondary,
+      "textLink.foreground"      : color.accent.fg,
+      "textLink.activeForeground": color.accent.fg,
+      "textBlockQuote.background": color.canvas.inset,
+      "textBlockQuote.border"    : color.border.default,
+      "textCodeBlock.background" : color.neutral.muted,
+      "textPreformat.foreground" : color.fg.muted,
+      "textSeparator.foreground" : color.border.muted,
+
+      "icon.foreground"           : color.fg.muted,
+      "keybindingLabel.foreground": color.fg.default,
 
       "button.background"     : color.btn.primary.bg,
       "button.foreground"     : color.btn.primary.text,
       "button.hoverBackground": color.btn.primary.hoverBg,
-      
+
       "button.secondaryBackground"     : color.btn.activeBg,
       "button.secondaryForeground"     : color.btn.text,
       "button.secondaryHoverBackground": color.btn.hoverBg,
 
-      "checkbox.background": color.bg.tertiary,
-      "checkbox.border"    : color.border.primary,
+      "checkbox.background": color.canvas.subtle,
+      "checkbox.border"    : color.border.default,
 
-      "dropdown.background"    : color.bg.overlay,
-      "dropdown.border"        : color.border.primary,
-      "dropdown.foreground"    : color.text.primary,
-      "dropdown.listBackground": color.bg.overlay,
+      "dropdown.background"    : color.canvas.overlay,
+      "dropdown.border"        : color.border.default,
+      "dropdown.foreground"    : color.fg.default,
+      "dropdown.listBackground": color.canvas.overlay,
 
-      "input.background"           : color.input.bg,
-      "input.border"               : color.input.border,
-      "input.foreground"           : color.text.primary,
-      "input.placeholderForeground": color.text.placeholder,
+      "input.background"           : color.canvas.default,
+      "input.border"               : color.border.default,
+      "input.foreground"           : color.fg.default,
+      "input.placeholderForeground": color.fg.subtle,
 
-      "badge.foreground": themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
-      "badge.background": themes({ light: scale.blue[1], dark: scale.blue[7], dimmed: scale.blue[7] }),
+      "badge.foreground": color.fg.onEmphasis,
+      "badge.background": color.accent.emphasis,
 
-      "progressBar.background": themes({ light: scale.blue[4], dark: scale.blue[5], dimmed: scale.blue[5] }),
+      "progressBar.background": color.accent.emphasis,
 
-      "titleBar.activeForeground"  : color.text.secondary,
-      "titleBar.activeBackground"  : color.bg.canvas,
-      "titleBar.inactiveForeground": color.text.tertiary,
-      "titleBar.inactiveBackground": color.bg.canvasInset,
-      "titleBar.border"            : color.border.primary,
+      "titleBar.activeForeground"  : color.fg.muted,
+      "titleBar.activeBackground"  : color.canvas.default,
+      "titleBar.inactiveForeground": color.fg.muted,
+      "titleBar.inactiveBackground": color.canvas.inset,
+      "titleBar.border"            : color.border.default,
 
-      "activityBar.foreground"        : color.text.primary,
-      "activityBar.inactiveForeground": color.text.tertiary,
-      "activityBar.background"        : color.bg.canvas,
-      "activityBarBadge.foreground"   : scale.white,
-      "activityBarBadge.background"   : themes({ light: scale.blue[4], dark: scale.blue[5], dimmed: scale.blue[5] }),
-      "activityBar.activeBorder"      : color.underlinenav.borderActive,
-      "activityBar.border"            : color.border.primary,
+      "activityBar.foreground"        : color.fg.default,
+      "activityBar.inactiveForeground": color.fg.muted,
+      "activityBar.background"        : color.canvas.default,
+      "activityBarBadge.foreground"   : color.fg.onEmphasis,
+      "activityBarBadge.background"   : color.accent.emphasis,
+      "activityBar.activeBorder"      : color.primer.border.active,
+      "activityBar.border"            : color.border.default,
 
-      "sideBar.foreground"             : color.text.primary,
-      "sideBar.background"             : color.bg.canvasInset,
-      "sideBar.border"                 : color.border.primary,
-      "sideBarTitle.foreground"        : color.text.primary,
-      "sideBarSectionHeader.foreground": color.text.primary,
-      "sideBarSectionHeader.background": color.bg.canvasInset,
-      "sideBarSectionHeader.border"    : color.border.primary,
+      "sideBar.foreground"             : color.fg.default,
+      "sideBar.background"             : color.canvas.inset,
+      "sideBar.border"                 : color.border.default,
+      "sideBarTitle.foreground"        : color.fg.default,
+      "sideBarSectionHeader.foreground": color.fg.default,
+      "sideBarSectionHeader.background": color.canvas.inset,
+      "sideBarSectionHeader.border"    : color.border.default,
 
-      "list.hoverForeground"            : color.text.primary,
-      "list.inactiveSelectionForeground": color.text.primary,
-      "list.activeSelectionForeground"  : color.text.primary,
-      "list.hoverBackground"            : themes({ light: "#ebf0f4", dark: scale.gray[8], dimmed: scale.gray[8] }),
-      "list.inactiveSelectionBackground": themes({ light: "#e8eaed", dark: scale.gray[8], dimmed: scale.gray[8] }),
-      "list.activeSelectionBackground"  : themes({ light: "#e2e5e9", dark: scale.gray[7], dimmed: scale.gray[7] }),
-      "list.focusForeground"            : themes({ light: scale.blue[9], dark: scale.gray[0], dimmed: scale.gray[0] }),
-      "list.focusBackground"            : themes({ light: "#cce5ff", dark: scale.gray[7], dimmed: scale.gray[7] }),
-      "list.inactiveFocusBackground"    : themes({ light: scale.blue[1], dark: scale.gray[8], dimmed: scale.gray[8] }),
-      "list.highlightForeground"        : themes({ light: scale.blue[5], dark: scale.blue[4], dimmed: scale.blue[4] }),
+      "list.hoverForeground"            : color.fg.default,
+      "list.inactiveSelectionForeground": color.fg.default,
+      "list.activeSelectionForeground"  : color.fg.default,
+      "list.hoverBackground"            : color.neutral.subtle,
+      "list.inactiveSelectionBackground": color.neutral.muted,
+      "list.activeSelectionBackground"  : color.neutral.muted,
+      "list.focusForeground"            : color.fg.default,
+      "list.focusBackground"            : color.accent.subtle,
+      "list.inactiveFocusBackground"    : color.accent.subtle,
+      "list.highlightForeground"        : color.accent.fg,
 
-      "tree.indentGuidesStroke": color.border.secondary,
+      "tree.indentGuidesStroke": color.border.muted,
 
-      "notificationCenterHeader.foreground": themes({ light: scale.gray[5], dark: scale.gray[4], dimmed: scale.gray[4] }),
-      "notificationCenterHeader.background": themes({ light: scale.gray[2], dark: scale.gray[9], dimmed: scale.gray[9] }),
-      "notifications.foreground"           : color.text.secondary,
-      "notifications.background"           : themes({ light: scale.gray[0], dark: scale.gray[8], dimmed: scale.gray[8] }),
-      "notifications.border"               : color.border.primary,
-      "notificationsErrorIcon.foreground"  : themes({ light: scale.red[5], dark: scale.red[4], dimmed: scale.red[4] }),
-      "notificationsWarningIcon.foreground": themes({ light: scale.orange[6], dark: scale.orange[3], dimmed: scale.orange[3] }),
-      "notificationsInfoIcon.foreground"   : themes({ light: scale.blue[6], dark: scale.blue[3], dimmed: scale.blue[3] }),
+      "notificationCenterHeader.foreground": color.fg.muted,
+      "notificationCenterHeader.background": color.canvas.subtle,
+      "notifications.foreground"           : color.fg.default,
+      "notifications.background"           : color.canvas.overlay,
+      "notifications.border"               : color.border.default,
+      "notificationsErrorIcon.foreground"  : color.danger.fg,
+      "notificationsWarningIcon.foreground": color.attention.fg,
+      "notificationsInfoIcon.foreground"   : color.accent.fg,
 
-      "pickerGroup.border"    : themes({ light: scale.gray[2], dark: scale.gray[7], dimmed: scale.gray[7] }),
-      "pickerGroup.foreground": color.text.secondary,
-      "quickInput.background" : themes({ light: scale.gray[0], dark: scale.gray[9], dimmed: scale.gray[9] }),
-      "quickInput.foreground" : color.text.primary,
+      "pickerGroup.border"    : color.border.default,
+      "pickerGroup.foreground": color.fg.muted,
+      "quickInput.background" : color.canvas.overlay,
+      "quickInput.foreground" : color.fg.default,
 
-      "statusBar.foreground"             : color.text.secondary,
-      "statusBar.background"             : color.bg.canvas,
-      "statusBar.border"                 : color.border.primary,
-      "statusBar.noFolderBackground"     : color.bg.canvas,
-      "statusBar.debuggingBackground"    : color.bg.dangerInverse,
-      "statusBar.debuggingForeground"    : scale.white,
-      "statusBarItem.prominentBackground": color.bg.tertiary,
+      "statusBar.foreground"             : color.fg.muted,
+      "statusBar.background"             : color.canvas.default,
+      "statusBar.border"                 : color.border.default,
+      "statusBar.noFolderBackground"     : color.canvas.default,
+      "statusBar.debuggingBackground"    : color.danger.emphasis,
+      "statusBar.debuggingForeground"    : color.fg.onEmphasis,
+      "statusBarItem.prominentBackground": color.canvas.subtle,
+      "statusBarItem.remoteForeground"   : color.fg.muted,
+      "statusBarItem.remoteBackground"   : color.canvas.default,
 
-      "editorGroupHeader.tabsBackground": color.bg.canvasInset,
-      "editorGroupHeader.tabsBorder"    : color.border.primary,
-      "editorGroup.border"              : color.border.primary,
+      "editorGroupHeader.tabsBackground": color.canvas.inset,
+      "editorGroupHeader.tabsBorder"    : color.border.default,
+      "editorGroup.border"              : color.border.default,
 
-      "tab.activeForeground"        : color.text.primary,
-      "tab.inactiveForeground"      : color.text.tertiary,
-      "tab.inactiveBackground"      : color.bg.canvasInset,
-      "tab.activeBackground"        : color.bg.canvas,
-      "tab.hoverBackground"         : color.bg.canvas,
-      "tab.unfocusedHoverBackground": color.state.hover.secondaryBg,
-      "tab.border"                  : color.border.primary,
-      "tab.unfocusedActiveBorderTop": color.border.primary,
-      "tab.activeBorder"            : color.bg.canvas,
-      "tab.unfocusedActiveBorder"   : color.bg.canvas,
-      "tab.activeBorderTop"         : color.underlinenav.borderActive,
+      "tab.activeForeground"        : color.fg.default,
+      "tab.inactiveForeground"      : color.fg.muted,
+      "tab.inactiveBackground"      : color.canvas.inset,
+      "tab.activeBackground"        : color.canvas.default,
+      "tab.hoverBackground"         : color.canvas.default,
+      "tab.unfocusedHoverBackground": color.neutral.subtle,
+      "tab.border"                  : color.border.default,
+      "tab.unfocusedActiveBorderTop": color.border.default,
+      "tab.activeBorder"            : color.canvas.default,
+      "tab.unfocusedActiveBorder"   : color.canvas.default,
+      "tab.activeBorderTop"         : color.primer.border.active,
 
-      "breadcrumb.foreground"               : color.text.tertiary,
-      "breadcrumb.focusForeground"          : color.text.primary,
-      "breadcrumb.activeSelectionForeground": color.text.secondary,
-      "breadcrumbPicker.background"         : color.bg.overlay,
+      "breadcrumb.foreground"               : color.fg.muted,
+      "breadcrumb.focusForeground"          : color.fg.default,
+      "breadcrumb.activeSelectionForeground": color.fg.muted,
+      "breadcrumbPicker.background"         : color.canvas.overlay,
 
-      "editor.foreground"                 : color.text.primary,
-      "editor.background"                 : color.bg.canvas,
-      "editorWidget.background"           : color.bg.overlay,
-      "editor.foldBackground"             : chroma(scale.gray[4]).alpha(0.1).hex(), // needs opacity
+      "editor.foreground"                 : color.fg.default,
+      "editor.background"                 : color.canvas.default,
+      "editorWidget.background"           : color.canvas.overlay,
+      "editor.foldBackground"             : alpha(color.neutral.emphasis, 0.1),
       "editor.lineHighlightBackground"    : color.codemirror.activelineBg,
+      "editor.lineHighlightBorder"        : onlyDarkHighContrast(color.accent.fg),
       "editorLineNumber.foreground"       : color.codemirror.linenumberText,
-      "editorLineNumber.activeForeground" : color.text.primary,
-      "editorIndentGuide.background"      : color.border.secondary,
-      "editorIndentGuide.activeBackground": color.border.primary,
-      "editorWhitespace.foreground"       : themes({ light: scale.gray[3], dark: scale.gray[5], dimmed: scale.gray[5] }),
-      "editorCursor.foreground"           : themes({ light: scale.blue[7], dark: scale.blue[2], dimmed: scale.blue[2] }),
+      "editorLineNumber.activeForeground" : color.fg.default,
+      "editorIndentGuide.background"      : color.border.muted,
+      "editorIndentGuide.activeBackground": color.border.default,
+      "editorWhitespace.foreground"       : color.fg.subtle,
+      "editorCursor.foreground"           : color.accent.fg,
 
-      "editor.findMatchBackground"          : themes({ light: scale.yellow[4], dark: "#ffd33d44", dimmed: "#ffd33d44" }),
-      "editor.findMatchHighlightBackground" : themes({ light: "#ffdf5d66", dark: "#ffd33d22", dimmed: "#ffd33d22" }),
-      "editor.linkedEditingBackground"      : themes({ light: "#0366d611", dark: "#3392FF22", dimmed: "#3392FF22" }),
-      "editor.inactiveSelectionBackground"  : themes({ light: "#0366d611", dark: "#3392FF22", dimmed: "#3392FF22" }),
-      "editor.selectionBackground"          : themes({ light: "#0366d625", dark: "#3392FF44", dimmed: "#3392FF44" }),
-      "editor.selectionHighlightBackground" : themes({ light: "#34d05840", dark: "#17E5E633", dimmed: "#17E5E633" }),
-      "editor.selectionHighlightBorder"     : themes({ light: "#34d05800", dark: "#17E5E600", dimmed: "#17E5E600" }),
-      "editor.wordHighlightBackground"      : themes({ light: "#34d05800", dark: "#17E5E600", dimmed: "#17E5E600" }),
-      "editor.wordHighlightStrongBackground": themes({ light: "#34d05800", dark: "#17E5E600", dimmed: "#17E5E600" }),
-      "editor.wordHighlightBorder"          : themes({ light: "#24943e99", dark: "#17E5E699", dimmed: "#17E5E699" }),
-      "editor.wordHighlightStrongBorder"    : themes({ light: "#24943e50", dark: "#17E5E666", dimmed: "#17E5E666" }),
-      "editorBracketMatch.background"       : themes({ light: "#34d05840", dark: "#17E5E650", dimmed: "#17E5E650" }),
-      "editorBracketMatch.border"           : themes({ light: "#34d05800", dark: "#17E5E600", dimmed: "#17E5E600" }),
+      "editor.findMatchBackground"            : color.attention.emphasis,
+      "editor.findMatchHighlightBackground"   : alpha(scale.yellow[1], 0.5),
+      "editor.linkedEditingBackground"        : alpha(color.accent.fg, 0.07),
+      "editor.inactiveSelectionBackground"    : alpha(color.accent.fg, 0.07),
+      "editor.selectionBackground"            : alpha(color.accent.fg, 0.2),
+      "editor.selectionHighlightBackground"   : alpha(scale.green[3], 0.25),
+      "editor.wordHighlightBackground"        : alpha(color.neutral.subtle, 0.5),
+      "editor.wordHighlightBorder"            : alpha(color.neutral.muted, 0.6),
+      "editor.wordHighlightStrongBackground"  : alpha(color.neutral.muted, 0.3),
+      "editor.wordHighlightStrongBorder"      : alpha(color.neutral.muted, 0.6),
+      "editorBracketMatch.background"         : alpha(scale.green[3], 0.25),
+      "editorBracketMatch.border"             : alpha(scale.green[3], 0.6),
 
-      "editorGutter.modifiedBackground": color.diff.change.border,
-      "editorGutter.addedBackground"   : color.diff.addition.border,
-      "editorGutter.deletedBackground" : color.diff.deletion.border,
+      "editorInlayHint.background": alpha(scale.gray[3], 0.2),
+      "editorInlayHint.foreground": color.fg.muted,
+      "editorInlayHint.typeBackground": alpha(scale.gray[3], 0.2),
+      "editorInlayHint.typeForeground": color.fg.muted,
+      "editorInlayHint.paramBackground": alpha(scale.gray[3], 0.2),
+      "editorInlayHint.paramForeground": color.fg.muted,
 
-      "diffEditor.insertedTextBackground": themes({ light: chroma(scale.green[3]).alpha(0.2).hex(), dark: chroma(color.diff.addition.bg).hex(), dimmed: chroma(color.diff.addition.bg).hex() }),
-      "diffEditor.removedTextBackground" : themes({ light: chroma(scale.red[3]  ).alpha(0.15).hex(), dark: chroma(color.diff.deletion.bg).hex(), dimmed: chroma(color.diff.deletion.bg).hex() }),
+      "editorGutter.modifiedBackground": color.attention.muted,
+      "editorGutter.addedBackground"   : color.success.muted,
+      "editorGutter.deletedBackground" : color.danger.muted,
 
-      "scrollbar.shadow"                : themes({ light: "#6a737d33", dark: "#0008", dimmed: "#0008" }),
-      "scrollbarSlider.background"      : themes({ light: "#959da533", dark: "#484F5833", dimmed: "#484F5833" }),
-      "scrollbarSlider.hoverBackground" : themes({ light: "#959da544", dark: "#484F5844", dimmed: "#484F5844" }),
-      "scrollbarSlider.activeBackground": themes({ light: "#959da588", dark: "#484F5888", dimmed: "#484F5888" }),
-      "editorOverviewRuler.border"      : themes({ light: scale.white, dark: scale.black, dimmed: scale.black }),
+      "diffEditor.insertedLineBackground": lightDark(alpha(scale.green[1], 0.3), alpha(scale.green[5], 0.2)),
+      "diffEditor.insertedTextBackground": lightDark(alpha(scale.green[2], 0.4), alpha(scale.green[5], 0.3)),
+      "diffEditor.removedLineBackground" : lightDark(alpha(scale.red[1], 0.3), alpha(scale.red[5], 0.2)),
+      "diffEditor.removedTextBackground" : lightDark(alpha(scale.red[2], 0.4), alpha(scale.red[5], 0.3)),
 
-      "panel.background"             : color.bg.canvasInset,
-      "panel.border"                 : color.border.primary,
-      "panelTitle.activeBorder"      : color.underlinenav.borderActive,
-      "panelTitle.activeForeground"  : color.text.primary,
-      "panelTitle.inactiveForeground": color.text.tertiary,
-      "panelInput.border"            : color.border.primary,
+      "scrollbar.shadow"                  : alpha(scale.gray[5], 0.2),
+      "scrollbarSlider.background"        : alpha(scale.gray[4], 0.2),
+      "scrollbarSlider.hoverBackground"   : alpha(scale.gray[4], 0.27),
+      "scrollbarSlider.activeBackground"  : alpha(scale.gray[4], 0.53),
+      "editorOverviewRuler.border"        : lightDark(scale.white, scale.black),
 
-      "terminal.foreground": color.text.secondary,
+      "panel.background"               : color.canvas.inset,
+      "panel.border"                   : color.border.default,
+      "panelTitle.activeBorder"        : color.primer.border.active,
+      "panelTitle.activeForeground"    : color.fg.default,
+      "panelTitle.inactiveForeground"  : color.fg.muted,
+      "panelInput.border"              : color.border.default,
+
+      "terminal.foreground": color.fg.muted,
       'terminal.ansiBlack': color.ansi.black,
       'terminal.ansiRed': color.ansi.red,
       'terminal.ansiGreen': color.ansi.green,
@@ -201,25 +232,25 @@ function getTheme({ theme, name }) {
       'terminal.ansiBrightCyan': color.ansi.cyanBright,
       'terminal.ansiBrightWhite': color.ansi.whiteBright,
 
-      "gitDecoration.addedResourceForeground"      : color.diff.addition.text,
-      "gitDecoration.modifiedResourceForeground"   : color.diff.change.text,
-      "gitDecoration.deletedResourceForeground"    : color.diff.deletion.text,
-      "gitDecoration.untrackedResourceForeground"  : color.diff.addition.text,
-      "gitDecoration.ignoredResourceForeground"    : themes({ light: scale.gray[4], dark: color.text.disabled, dimmed: color.text.disabled }),
-      "gitDecoration.conflictingResourceForeground": color.text.warning,
-      "gitDecoration.submoduleResourceForeground"  : color.text.secondary,
+      "gitDecoration.addedResourceForeground"      : color.success.fg,
+      "gitDecoration.modifiedResourceForeground"   : color.attention.fg,
+      "gitDecoration.deletedResourceForeground"    : color.danger.fg,
+      "gitDecoration.untrackedResourceForeground"  : color.success.fg,
+      "gitDecoration.ignoredResourceForeground"    : color.fg.subtle,
+      "gitDecoration.conflictingResourceForeground": color.severe.fg,
+      "gitDecoration.submoduleResourceForeground"  : color.fg.muted,
 
-      "debugToolBar.background"                    : color.bg.overlay,
-      "editor.stackFrameHighlightBackground"       : themes({ light: "#ffd33d33", dark: "#D2992225", dimmed: "#C6902625" }), // needs opacity (yellow)
-      "editor.focusedStackFrameHighlightBackground": themes({ light: "#28a74525", dark: "#3FB95025", dimmed: "#2b6a3033" }), // needs opacity (green)
+      "debugToolBar.background"                    : color.canvas.overlay,
+      "editor.stackFrameHighlightBackground"       : color.attention.muted,
+      "editor.focusedStackFrameHighlightBackground": color.success.muted,
 
-      "peekViewEditor.matchHighlightBackground": themes({ dark: "#ffd33d33", dimmed: "#ffd33d33" }),
-      "peekViewResult.matchHighlightBackground": themes({ dark: "#ffd33d33", dimmed: "#ffd33d33" }),
-      "peekViewEditor.background"              : themes({ dark: "#0d111788", dimmed: "#0d111788" }),
-      "peekViewResult.background"              : themes({ dark: scale.gray[9], dimmed: scale.gray[9] }),
+      "peekViewEditor.matchHighlightBackground": onlyDark(color.attention.muted),
+      "peekViewResult.matchHighlightBackground": onlyDark(color.attention.muted),
+      "peekViewEditor.background"              : onlyDark(color.neutral.subtle),
+      "peekViewResult.background"              : onlyDark(scale.gray[9]),
 
-      "settings.headerForeground"        : color.text.secondary,
-      "settings.modifiedItemIndicator"   : color.diff.change.border,
+      "settings.headerForeground"        : color.fg.muted,
+      "settings.modifiedItemIndicator"   : color.attention.muted,
       "welcomePage.buttonBackground"     : color.btn.bg,
       "welcomePage.buttonHoverBackground": color.btn.hoverBg,
     },
@@ -228,7 +259,7 @@ function getTheme({ theme, name }) {
       {
         scope: ["comment", "punctuation.definition.comment", "string.comment"],
         settings: {
-          foreground: themes({ light: scale.gray[5], dark: scale.gray[3], dimmed: scale.gray[3] }),
+          foreground: lightDark(scale.gray[5], scale.gray[3])
         },
       },
       {
@@ -236,11 +267,12 @@ function getTheme({ theme, name }) {
           "constant",
           "entity.name.constant",
           "variable.other.constant",
+          "variable.other.enummember",
           "variable.language",
           "entity",
         ],
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
@@ -250,7 +282,7 @@ function getTheme({ theme, name }) {
           "meta.definition.variable"
         ],
         settings: {
-          foreground: themes({ light: scale.orange[6], dark: scale.orange[2], dimmed: scale.orange[2] }),
+          foreground: lightDark(scale.orange[6], scale.orange[2])
         },
       },
       {
@@ -264,13 +296,13 @@ function getTheme({ theme, name }) {
           "meta.embedded.expression"
         ],
         settings: {
-          foreground: color.text.primary,
+          foreground: color.fg.default,
         },
       },
       {
         "scope": "entity.name.function",
         "settings": {
-          foreground: themes({ light: scale.purple[5], dark: scale.purple[2], dimmed: scale.purple[2] }),
+          foreground: lightDark(scale.purple[5], scale.purple[2])
         }
       },
       {
@@ -279,19 +311,19 @@ function getTheme({ theme, name }) {
           "support.class.component"
         ],
         settings: {
-          foreground: themes({ light: scale.green[6], dark: scale.green[1], dimmed: scale.green[1] }),
+          foreground: lightDark(scale.green[6], scale.green[1])
         },
       },
       {
         scope: "keyword",
         settings: {
-          foreground: themes({ light: scale.red[5], dark: scale.red[3], dimmed: scale.red[3] }),
+          foreground: lightDark(scale.red[5], scale.red[3])
         },
       },
       {
         scope: ["storage", "storage.type"],
         settings: {
-          foreground: themes({ light: scale.red[5], dark: scale.red[3], dimmed: scale.red[3] }),
+          foreground: lightDark(scale.red[5], scale.red[3])
         },
       },
       {
@@ -301,7 +333,7 @@ function getTheme({ theme, name }) {
           "storage.type.java",
         ],
         settings: {
-          foreground: color.text.primary,
+          foreground: color.fg.default,
         },
       },
       {
@@ -311,92 +343,92 @@ function getTheme({ theme, name }) {
           "string punctuation.section.embedded source",
         ],
         settings: {
-          foreground: themes({ light: scale.blue[8], dark: scale.blue[1], dimmed: scale.blue[1] }),
+          foreground: lightDark(scale.blue[8], scale.blue[1])
         },
       },
       {
         scope: "support",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "meta.property-name",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "variable",
         settings: {
-          foreground: themes({ light: scale.orange[6], dark: scale.orange[2], dimmed: scale.orange[2] }),
+          foreground: lightDark(scale.orange[6], scale.orange[2])
         },
       },
       {
         scope: "variable.other",
         settings: {
-          foreground: color.text.primary,
+          foreground: color.fg.default,
         },
       },
       {
         scope: "invalid.broken",
         settings: {
           fontStyle: "italic",
-          foreground: themes({ light: scale.red[7], dark: scale.red[2], dimmed: scale.red[2] }),
+          foreground: lightDark(scale.red[7], scale.red[2])
         },
       },
       {
         scope: "invalid.deprecated",
         settings: {
           fontStyle: "italic",
-          foreground: themes({ light: scale.red[7], dark: scale.red[2], dimmed: scale.red[2] }),
+          foreground: lightDark(scale.red[7], scale.red[2])
         },
       },
       {
         scope: "invalid.illegal",
         settings: {
           fontStyle: "italic",
-          foreground: themes({ light: scale.red[7], dark: scale.red[2], dimmed: scale.red[2] }),
+          foreground: lightDark(scale.red[7], scale.red[2])
         },
       },
       {
         scope: "invalid.unimplemented",
         settings: {
           fontStyle: "italic",
-          foreground: themes({ light: scale.red[7], dark: scale.red[2], dimmed: scale.red[2] }),
+          foreground: lightDark(scale.red[7], scale.red[2])
         },
       },
       {
         scope: "carriage-return",
         settings: {
           fontStyle: "italic underline",
-          background: themes({ light: scale.red[5], dark: scale.red[3], dimmed: scale.red[3] }),
-          foreground: themes({ light: scale.gray[0], dark: scale.gray[9], dimmed: scale.gray[9] }),
+          background: lightDark(scale.red[5], scale.red[3]),
+          foreground: lightDark(scale.gray[0], scale.gray[0]),
           content: "^M",
         },
       },
       {
         scope: "message.error",
         settings: {
-          foreground: themes({ light: scale.red[7], dark: scale.red[2], dimmed: scale.red[2] }),
+          foreground: lightDark(scale.red[7], scale.red[2])
         },
       },
       {
         scope: "string source",
         settings: {
-          foreground: color.text.primary,
+          foreground: color.fg.default,
         },
       },
       {
         scope: "string variable",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: ["source.regexp", "string.regexp"],
         settings: {
-          foreground: themes({ light: scale.blue[8], dark: scale.blue[1], dimmed: scale.blue[1] }),
+          foreground: lightDark(scale.blue[8], scale.blue[1])
         },
       },
       {
@@ -407,71 +439,77 @@ function getTheme({ theme, name }) {
           "string.regexp string.regexp.arbitrary-repitition",
         ],
         settings: {
-          foreground: themes({ light: scale.blue[8], dark: scale.blue[1], dimmed: scale.blue[1] }),
+          foreground: lightDark(scale.blue[8], scale.blue[1])
         },
       },
       {
         scope: "string.regexp constant.character.escape",
         settings: {
           fontStyle: "bold",
-          foreground: themes({ light: scale.green[6], dark: scale.green[1], dimmed: scale.green[1] }),
+          foreground: lightDark(scale.green[6], scale.green[1])
         },
       },
       {
         scope: "support.constant",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "support.variable",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "meta.module-reference",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "punctuation.definition.list.begin.markdown",
         settings: {
-          foreground: themes({ light: scale.orange[6], dark: scale.orange[2], dimmed: scale.orange[2] }),
+          foreground: lightDark(scale.orange[6], scale.orange[2])
         },
       },
       {
         scope: ["markup.heading", "markup.heading entity.name"],
         settings: {
           fontStyle: "bold",
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "markup.quote",
         settings: {
-          foreground: themes({ light: scale.green[6], dark: scale.green[1], dimmed: scale.green[1] }),
+          foreground: lightDark(scale.green[6], scale.green[1])
         },
       },
       {
         scope: "markup.italic",
         settings: {
           fontStyle: "italic",
-          foreground: color.text.primary,
+          foreground: color.fg.default,
         },
       },
       {
         scope: "markup.bold",
         settings: {
           fontStyle: "bold",
-          foreground: color.text.primary,
+          foreground: color.fg.default,
         },
       },
       {
-        scope: "markup.raw",
+        scope: ["markup.underline"],
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          fontStyle: "underline",
+        },
+      },
+      {
+        scope: "markup.inline.raw",
+        settings: {
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
@@ -481,8 +519,8 @@ function getTheme({ theme, name }) {
           "punctuation.definition.deleted",
         ],
         settings: {
-          background: themes({ light: scale.red[0], dark: scale.red[9], dimmed: scale.red[9] }),
-          foreground: themes({ light: scale.red[7], dark: scale.red[2], dimmed: scale.red[2] }),
+          background: lightDark(scale.red[0], scale.red[9]),
+          foreground: lightDark(scale.red[7], scale.red[2])
         },
       },
       {
@@ -492,48 +530,48 @@ function getTheme({ theme, name }) {
           "punctuation.definition.inserted",
         ],
         settings: {
-          background: themes({ light: scale.green[0], dark: scale.green[9], dimmed: scale.green[9] }),
-          foreground: themes({ light: scale.green[6], dark: scale.green[1], dimmed: scale.green[1] }),
+          background: lightDark(scale.green[0], scale.green[9]),
+          foreground: lightDark(scale.green[6], scale.green[1])
         },
       },
       {
         scope: ["markup.changed", "punctuation.definition.changed"],
         settings: {
-          background: themes({ light: scale.orange[1], dark: scale.orange[8], dimmed: scale.orange[8] }),
-          foreground: themes({ light: scale.orange[6], dark: scale.orange[2], dimmed: scale.orange[2] }),
+          background: lightDark(scale.orange[1], scale.orange[8]),
+          foreground: lightDark(scale.orange[6], scale.orange[2])
         },
       },
       {
         scope: ["markup.ignored", "markup.untracked"],
         settings: {
-          foreground: themes({ light: scale.gray[1], dark: scale.gray[8], dimmed: scale.gray[8] }),
-          background: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.gray[1], scale.gray[8]),
+          background: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "meta.diff.range",
         settings: {
-          foreground: themes({ light: scale.purple[5], dark: scale.purple[2], dimmed: scale.purple[2] }),
+          foreground: lightDark(scale.purple[5], scale.purple[2]),
           fontStyle: "bold",
         },
       },
       {
         scope: "meta.diff.header",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "meta.separator",
         settings: {
           fontStyle: "bold",
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
         scope: "meta.output",
         settings: {
-          foreground: themes({ light: scale.blue[6], dark: scale.blue[2], dimmed: scale.blue[2] }),
+          foreground: lightDark(scale.blue[6], scale.blue[2])
         },
       },
       {
@@ -546,19 +584,19 @@ function getTheme({ theme, name }) {
           "brackethighlighter.quote",
         ],
         settings: {
-          foreground: themes({ light: scale.gray[6], dark: scale.gray[3], dimmed: scale.gray[3] }),
+          foreground: lightDark(scale.gray[6], scale.gray[3])
         },
       },
       {
         scope: "brackethighlighter.unmatched",
         settings: {
-          foreground: themes({ light: scale.red[7], dark: scale.red[2], dimmed: scale.red[2] }),
+          foreground: lightDark(scale.red[7], scale.red[2])
         },
       },
       {
         scope: ["constant.other.reference.link", "string.other.link"],
         settings: {
-          foreground: themes({ light: scale.blue[8], dark: scale.blue[1], dimmed: scale.blue[1] }),
+          foreground: lightDark(scale.blue[8], scale.blue[1]),
           fontStyle: "underline",
         },
       },
@@ -576,5 +614,25 @@ function getTheme({ theme, name }) {
     ],
   };
 }
+
+// Convert to hex
+// VS Code doesn't support other formats like hsl, rgba etc.
+
+function changeColorToHexAlphas(obj) {
+  if (typeof obj === 'object') {
+    for (var keys in obj) {
+      if (typeof obj[keys] === 'object') {
+        changeColorToHexAlphas(obj[keys])
+      } else {
+        let keyValue = obj[keys]
+        if(chroma.valid(keyValue)){
+          obj[keys] = chroma(keyValue).hex();
+        }
+      }
+    }
+  }
+  return obj;
+}
+
 
 module.exports = getTheme;