+++ CHANGELOG.md
@@ -1,3 +1,93 @@
+# github-vscode-theme
+
+## 6.1.0
+
+### Minor Changes
+
+- [#212](https://github.com/primer/github-vscode-theme/pull/212) [`f03ca46`](https://github.com/primer/github-vscode-theme/commit/f03ca46b7ace7d98f8e6761dca56083e1ba54c3c) Thanks [@akiirui](https://github.com/akiirui)! - fix: add missing scope variable.other.enummember
+
+* [#193](https://github.com/primer/github-vscode-theme/pull/193) [`8bb0bc8`](https://github.com/primer/github-vscode-theme/commit/8bb0bc8a2150515ec5c6d9467335b6b827870ec5) Thanks [@serpilacar](https://github.com/serpilacar)! - Update markdown inline code scope
+
+- [#252](https://github.com/primer/github-vscode-theme/pull/252) [`706236c`](https://github.com/primer/github-vscode-theme/commit/706236cfaa6bdeb7ba3942d63eb03a82097fcc04) Thanks [@magic-akari](https://github.com/magic-akari)! - feat: add underline support
+
+### Patch Changes
+
+- [#268](https://github.com/primer/github-vscode-theme/pull/268) [`6bc9e1a`](https://github.com/primer/github-vscode-theme/commit/6bc9e1a21f3251b1064683e8d8fce4172a56e2a5) Thanks [@simurai](https://github.com/simurai)! - Update remote statusbar item
+
+## 6.0.0
+
+### Major Changes
+
+- [#240](https://github.com/primer/github-vscode-theme/pull/240) [`075feb5`](https://github.com/primer/github-vscode-theme/commit/075feb5c753b3ac3fc2029e98203be5134bcd8f7) Thanks [@maximedegreve](https://github.com/maximedegreve)! - # Light High Contrast Theme
+
+  GitHub Light High Contrast is now out of beta and available to everyone. 🎉
+
+  ![Light High Contrast Theme](https://user-images.githubusercontent.com/980622/151374445-39670cae-ab83-400c-8b7d-fe4dd36c7cc3.png)
+
+## 5.2.2
+
+### Patch Changes
+
+- [#244](https://github.com/primer/github-vscode-theme/pull/244) [`d318d0e`](https://github.com/primer/github-vscode-theme/commit/d318d0eb3aa3a48c0983c3cc7b245847643b046a) Thanks [@maximedegreve](https://github.com/maximedegreve)! - Easier code to manage + port values to new primitives
+
+## 5.2.1
+
+### Patch Changes
+
+- [#256](https://github.com/primer/github-vscode-theme/pull/256) [`6549f32`](https://github.com/primer/github-vscode-theme/commit/6549f32ee5e97bb01655d4de47672c15e58e06c9) Thanks [@simurai](https://github.com/simurai)! - Fix Dark Dimmed from loosing syntax highlighting
+
+## 5.2.0
+
+### Minor Changes
+
+- [#240](https://github.com/primer/github-vscode-theme/pull/240) [`075feb5`](https://github.com/primer/github-vscode-theme/commit/075feb5c753b3ac3fc2029e98203be5134bcd8f7) Thanks [@maximedegreve](https://github.com/maximedegreve)! - Light High Contrast Theme
+
+## 5.1.0
+
+### Minor Changes
+
+- [#225](https://github.com/primer/github-vscode-theme/pull/225) [`241d482`](https://github.com/primer/github-vscode-theme/commit/241d48297ea3a7c91d6bf3c61b5beffb13044f32) Thanks [@maximedegreve](https://github.com/maximedegreve)! - # Colorblind Light & Dark Theme
+  Adds a new GitHub Colorblind Light and Dark theme similar to the themes on github.com.
+
+  ![Colorblind](https://user-images.githubusercontent.com/980622/136975341-4d5eca55-40d6-4d9f-8b33-ba8acbc805e4.png)
+
+## 5.0.3
+
+### Patch Changes
+
+- [#234](https://github.com/primer/github-vscode-theme/pull/234) [`bbb3a83`](https://github.com/primer/github-vscode-theme/commit/bbb3a83f20be57106f6bf68bb06b7d64b2a58025) Thanks [@simurai](https://github.com/simurai)! - Update README
+
+## 5.0.2
+
+### Patch Changes
+
+- [#232](https://github.com/primer/github-vscode-theme/pull/232) [`027ef70`](https://github.com/primer/github-vscode-theme/commit/027ef7050ba360a76d7bdd252ff60568edc4343b) Thanks [@jonrohan](https://github.com/jonrohan)! - Testing another patch release
+
+## 5.0.1
+
+### Patch Changes
+
+- [#229](https://github.com/primer/github-vscode-theme/pull/229) [`4279f43`](https://github.com/primer/github-vscode-theme/commit/4279f431387df11544508f618ecc75c3739a95ed) Thanks [@simurai](https://github.com/simurai)! - Add note about publishing
+
+## 5.0.0
+
+### Major Changes
+
+- Add new "GitHub Dark High Contrast" theme [#203](https://github.com/primer/github-vscode-theme/pull/203).
+- Update colors from Primer Primitives [#203](https://github.com/primer/github-vscode-theme/pull/203).
+
+![GitHub VS Code theme](https://user-images.githubusercontent.com/378023/132220037-3cd3e777-55a6-445f-9a2e-da6020ebd78d.png)
+
+## 4.2.1
+
+- Fix importing v2 colors from primer/primitives
+
+## 4.2.0
+
+- Add `editorBracketHighlight` and ANSI to classic. [#205](https://github.com/primer/github-vscode-theme/pull/205).
+- Set `terminal.tab.activeBorder` based on editor tab [#189](https://github.com/primer/github-vscode-theme/pull/189).
+- More clearer for the difference between new theme and legacy theme [#191](https://github.com/primer/github-vscode-theme/pull/191).
+
 ## 4.1.1
 
 - Generate themes based on `4.1.0`.
@@ -40,7 +130,7 @@
 
 ## 3.0.0
 
-- Revert dark changes  [#123](https://github.com/primer/github-vscode-theme/pull/123). This should be the same as the `1.1.5` version.
+- Revert dark changes [#123](https://github.com/primer/github-vscode-theme/pull/123). This should be the same as the `1.1.5` version.
 
 ## 2.0.1